requirements.txt
tensorflow==2.5.0
fastapi
uvicorn
python-multipart
pillow
numpy